MessageManager Implemented
Messages root added to DB. Messages have an ID and under that only two members : User and Server. They are both strings. We only keep track of the last one, and really this is only to set off the event for value change communication between apps.

When a ticket is picked up which does not have the message id it is given one and that id is also added under the Messages root.

Examples of how the observer value for the UI should look are in the MessageManager.

Aslo minor bug fixes in ticket model and ticket manager.